move funcao para o store
