+++ src/pages/register/user/index.tsx
@@ -6,28 +6,14 @@ import TabDadosBasicos from './tabs/tabDadosBasicos';
 import TabDadosPerfil from './tabs/tabPerfil';
 import { useRegisterUser } from '@stores/register/user/store';
 import { useNotification } from '@contexts/notificationContext';
-import { Link, useNavigate } from 'react-router-dom';
-
+import { Link } from 'react-router-dom';
 import './style.css';
-import { registerUser } from 'routes/routesUser';
 
 const RegisterUser = () => {
   const [activeIndex, setActiveIndex] = useState(0);
-  const { user, setField, validateStep } = useRegisterUser();
+  const { user, setField, validateStep, finalizeRegister } = useRegisterUser();
   const [_, setSubmitted] = useState(false);
   const { showNotification } = useNotification();
-
-  const navigate = useNavigate();
-
-  const finalizeRegister = async () => {
-    try {
-      await registerUser(user);
-      showNotification('success', null, 'Usuário cadastrado com sucesso!');
-      navigate('/login');
-    } catch {
-      showNotification('error', null, 'Erro ao cadastrar usuário!');
-    }
-  };
 
   const nextStep = (e: any) => {
     if (validateStep(activeIndex)) {
@@ -72,7 +58,7 @@ const RegisterUser = () => {
             <div style={{ display: 'flex', justifyContent: 'center', gap: '10px' }}>
               {activeIndex !== 0 && <Button className="step-button" label="Voltar" onClick={(e) => stepBack(e)} />}
               {activeIndex < 1 && <Button className="step-button" label="Continuar" onClick={(e) => nextStep(e)} />}
-                {activeIndex === 1 && <Button className="step-button" label="Finalizar" onClick={(e) => finalizeRegister()} />}
+                {activeIndex === 1 && <Button className="step-button" label="Finalizar" onClick={finalizeRegister} />}
             </div>
             <p className="login" style={{ color: '#2F292A', textAlign: 'center' }}>
               Já tem uma conta? <Link to="/login">Entrar</Link>
